% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcqFunctionGSy.R
\name{AcqFunctionGSy}
\alias{AcqFunctionGSy}
\title{GSy Acquisition Function}
\arguments{
\item{surrogate}{(\code{NULL} | \link[mlr3mbo:SurrogateLearner]{mlr3mbo::SurrogateLearner})\cr
Surrogate used for mean predictions.}
}
\description{
Greedy-sampling-in-output-space acquisition function.

The score is the minimum absolute distance between a candidate prediction and
an already observed target value.

Creates a new GSy acquisition function.
}
\seealso{
Other Acquisition Function:
\code{\link{AcqFunctionDist}},
\code{\link{AcqFunctionDistGSx}},
\code{\link{AcqFunctionDistIDEAL}},
\code{\link{AcqFunctionDistIGS}},
\code{\link{SurrogateNull}}
}
\concept{Acquisition Function}
