% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_scaling.R
\name{affine_scale_apply}
\alias{affine_scale_apply}
\title{Apply affine scaling to new data.}
\usage{
affine_scale_apply(x, scaler)
}
\arguments{
\item{x}{(\code{matrix})\cr
Numeric matrix with same number of columns as original pool.}

\item{scaler}{(\code{list})\cr
List returned by \code{\link[=affine_scale_features]{affine_scale_features()}}.}
}
\value{
Scaled \code{matrix}.
}
\description{
Apply affine scaling to new data.
}
\keyword{internal}
