% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_scaling.R
\name{affine_scaler_from_bounds}
\alias{affine_scaler_from_bounds}
\title{Build affine scaler from explicit bounds.}
\usage{
affine_scaler_from_bounds(lower, upper)
}
\arguments{
\item{lower}{(\code{numeric})\cr
Vector of lower bounds.}

\item{upper}{(\code{numeric})\cr
Vector of upper bounds.}
}
\value{
Named \code{list} with entries \code{mins} (\code{numeric}) and \code{maxs} (\code{numeric})
}
\description{
Build affine scaler from explicit bounds.
}
\keyword{internal}
