% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_al_objectives.R
\name{alobj}
\alias{alobj}
\title{Syntactic Sugar Active-Learning Benchmark Objective Construction}
\usage{
alobj(.key, ...)
}
\arguments{
\item{.key}{(\code{character(1)})\cr
Key passed to \link{mlr_al_objectives}.}

\item{...}{(named \code{list()})\cr
Named arguments passed to the constructor, to the parameter set, or to
public fields. See \code{\link[mlr3misc:dictionary_sugar_get]{mlr3misc::dictionary_sugar_get()}} for details.}
}
\value{
\link[bbotk:Objective]{bbotk::Objective}.
}
\description{
Retrieves an objective from \link{mlr_al_objectives}.
}
\examples{
alobj("scalar")
}
