% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_al_objectives.R
\name{alobjs}
\alias{alobjs}
\title{Syntactic Sugar Active-Learning Benchmark Objectives Construction}
\usage{
alobjs(.keys, ...)
}
\arguments{
\item{.keys}{(\code{character()})\cr
Keys passed to \link{mlr_al_objectives}.}

\item{...}{(named \code{list()})\cr
Named arguments passed to the constructor, to the parameter set, or to
public fields. See \code{\link[mlr3misc:dictionary_sugar_get]{mlr3misc::dictionary_sugar_mget()}} for details.}
}
\value{
Named \code{list} of \link[bbotk:Objective]{bbotk::Objective} objects.
}
\description{
Retrieves multiple objectives from \link{mlr_al_objectives}.
}
\examples{
alobjs(c("scalar", "bell"))
}
