% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALProposer.R
\name{ALProposerPortfolio}
\alias{ALProposerPortfolio}
\title{Portfolio Active Learning Proposer}
\arguments{
\item{proposers}{(named \code{list()} of \link{ALProposer})\cr
Child proposers.}
}
\description{
Combines child proposers by querying them round-robin until enough unique
points have been proposed or all children fail to produce another point.

Creates a portfolio proposer.
}
\section{Fields}{

\describe{
\item{\code{proposers}}{(named \code{list()} of \link{ALProposer})
Child proposers.}
}}

