% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALProposer.R
\name{ALProposerPseudoLabel}
\alias{ALProposerPseudoLabel}
\title{Pseudo-Label Active Learning Proposer}
\arguments{
\item{label_surrogate_id}{(\code{character(1)})\cr
Label surrogate registry id.}
}
\description{
Builds a batch sequentially by adding pseudo-labels to a temporary archive
between proposals.

The pseudo-label values are computed by the \code{label_surrogate_id} surrogate.

Creates a pseudo-label proposer.
}
\section{Fields}{

\describe{
\item{\code{label_surrogate_id}}{(\code{character(1)})\cr
Label surrogate registry id.}
}}

