% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALScoreModifier.R
\name{ALScoreModifierLocalPenalization}
\alias{ALScoreModifierLocalPenalization}
\title{Local-Penalization Active Learning Score Modifier}
\arguments{
\item{bandwidth}{(\code{numeric(1)})\cr
Gower-distance radius in which utilities are penalized.}

\item{penalization}{(\code{numeric(1)})\cr
Amount subtracted from utilities. \code{Inf} removes nearby points.}
}
\description{
Penalizes candidates close to points already selected for the current batch.

Creates a local-penalization score modifier.
}
