% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_task_lce.R
\name{as_task_lce}
\alias{as_task_lce}
\alias{as_task_lce.data.frame}
\alias{as_task_lce.TaskLCE}
\alias{as_task_lce.DataBackend}
\title{Convert to Learning Curve Extrapolation Task}
\usage{
as_task_lce(x, ...)

\method{as_task_lce}{data.frame}(
  x,
  target,
  order,
  key = character(),
  direction = "minimize",
  id = deparse1(substitute(x)),
  ...
)

\method{as_task_lce}{TaskLCE}(x, clone = FALSE, ...)

\method{as_task_lce}{DataBackend}(
  x,
  target,
  order,
  key = character(),
  direction = "minimize",
  id = deparse1(substitute(x)),
  ...
)
}
\arguments{
\item{x}{(\code{any})\cr
Object to convert.}

\item{...}{\cr
Additional arguments passed to \link{TaskLCE}.}

\item{target}{(\code{character(1)})\cr
Name of the target column.}

\item{order}{(\code{character(1)})\cr
Name of the order column (evaluation count).}

\item{key}{(\code{character()})\cr
Optional key column name(s).}

\item{direction}{(\code{character(1)})\cr
\code{"minimize"} or \code{"maximize"}.}

\item{id}{(\code{character(1)})\cr
Task identifier. Defaults to the variable name of \code{x}.}

\item{clone}{(\code{logical(1)})\cr
Whether to clone the task.}
}
\value{
\link{TaskLCE}.
}
\description{
Convert an object to a \link{TaskLCE}.
}
