% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_objective.R
\name{assert_pool_objective_search_space}
\alias{assert_pool_objective_search_space}
\title{Assert Pool Objective Search-Space Compatibility}
\usage{
assert_pool_objective_search_space(objective, search_space)
}
\arguments{
\item{objective}{(\link[bbotk:Objective]{bbotk::Objective})\cr
Objective to evaluate.}

\item{search_space}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Search space used to propose points.}
}
\value{
The validated \code{search_space}, invisibly.
}
\description{
Validates search-space features that pool-backed objectives cannot support.
}
\keyword{internal}
