% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallbackForecastTracker.R
\name{celecx.forecast_tracker}
\alias{celecx.forecast_tracker}
\alias{CallbackForecastTracker}
\title{Forecast Tracker Callback}
\arguments{
\item{metrics_tracker}{(\link{MetricsTracker})
Tracker containing the metrics history.}

\item{forecast_tracker}{(\code{ForecastTracker})
Tracker to be updated.}
}
\description{
Updates a \code{ForecastTracker} after each evaluated batch.

Creates a new CallbackForecastTracker.
}
\details{
This callback expects that a \link{MetricsTracker} callback has already logged the
current batch in the same callback chain.
}
\section{Fields}{

\describe{
\item{\code{metrics_tracker}}{(\link{MetricsTracker})}

\item{\code{forecast_tracker}}{(\code{ForecastTracker})}
}}

\examples{
\dontrun{
metrics_tracker <- MetricsTracker$new(metrics = list(best_y = metric_best_y))
forecast_tracker <- ForecastTracker$new(
  extrapolator = CurveExtrapolatorParametric$new(),
  metric_name = "best_y"
)
callback <- clbk("celecx.forecast_tracker",
  metrics_tracker = metrics_tracker,
  forecast_tracker = forecast_tracker
)
}
}
