% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallbackMetricsTracker.R
\name{celecx.metrics_tracker}
\alias{celecx.metrics_tracker}
\alias{CallbackMetricsTracker}
\title{Metrics Tracker Callback}
\arguments{
\item{metrics_tracker}{(\link{MetricsTracker})\cr
Tracker used for logging.}
}
\description{
Logs metrics after each batch evaluation using a \link{MetricsTracker}.

This callback is meant to be passed as \code{callbacks} to a bbotk
\link[bbotk:OptimInstanceBatch]{bbotk::OptimInstanceBatch} / \link[bbotk:OptimInstanceAsync]{bbotk::OptimInstanceAsync} (or via the
\code{callbacks} argument of \code{oi()} / \code{bb_optimize()}).

Creates a new CallbackMetricsTracker.
}
\section{Fields}{

\describe{
\item{\code{metrics_tracker}}{(\link{MetricsTracker})\cr
Tracker that collects per-batch metrics.}
}}

\examples{
\dontrun{
tracker <- MetricsTracker$new()
callback <- clbk("celecx.metrics_tracker", metrics_tracker = tracker)
}
}
