% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchInstance.R
\name{choose_search_space}
\alias{choose_search_space}
\title{Choose Search Space}
\usage{
choose_search_space(objective, search_space)
}
\arguments{
\item{objective}{(\link[bbotk:Objective]{bbotk::Objective})\cr
The objective with a domain ParamSet.}

\item{search_space}{(\link[paradox:ParamSet]{paradox::ParamSet} | \code{NULL})\cr
Optional user-supplied search space.}
}
\value{
A \link[paradox:ParamSet]{paradox::ParamSet} to use as the search space.
}
\description{
Derives the search space from the objective domain and optional user-supplied
search space. Follows the same logic as bbotk's OptimInstanceBatch.
}
\keyword{internal}
