% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_al_distances.R
\name{clx_ald}
\alias{clx_ald}
\title{Syntactic Sugar Active Learning Distance Construction}
\usage{
clx_ald(.key, ...)
}
\arguments{
\item{.key}{(\code{character(1)})\cr
Key passed to \link{mlr_al_distances}.}

\item{...}{(named \code{list()})\cr
Named arguments passed to the constructor, to the parameter set, or to
public fields. See \code{\link[mlr3misc:dictionary_sugar_get]{mlr3misc::dictionary_sugar_get()}} for details.}
}
\value{
\link{ALDistance}.
}
\description{
This function complements \link{mlr_al_distances} with a construction helper in
the spirit of \code{mlr_sugar}.
}
\examples{
clx_ald("standardize")
}
