% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_al_distances.R
\name{clx_alds}
\alias{clx_alds}
\title{Syntactic Sugar Active Learning Distances Construction}
\usage{
clx_alds(.keys, ...)
}
\arguments{
\item{.keys}{(\code{character()})\cr
Keys passed to \link{mlr_al_distances}.}

\item{...}{(named \code{list()})\cr
Named arguments passed to the constructor, to the parameter set, or to
public fields. See \code{\link[mlr3misc:dictionary_sugar_mget]{mlr3misc::dictionary_sugar_mget()}} for details.}
}
\value{
Named \code{list} of \link{ALDistance} objects.
}
\description{
Retrieve multiple active-learning distances from \link{mlr_al_distances}.
}
\examples{
clx_alds(c("standardize", "affine"))
}
