% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_space_samplers.R
\name{clx_sps}
\alias{clx_sps}
\title{Syntactic Sugar Space Sampler Construction}
\usage{
clx_sps(.key, ...)
}
\arguments{
\item{.key}{(\code{character(1)})\cr
Key passed to \link{mlr_space_samplers}.}

\item{...}{(named \code{list()})\cr
Named arguments passed to the constructor, to the parameter set, or to
public fields. See \code{\link[mlr3misc:dictionary_sugar_get]{mlr3misc::dictionary_sugar_get()}} for details.}
}
\value{
\link{SpaceSampler}.
}
\description{
This function complements \link{mlr_space_samplers} with a construction helper in
the spirit of \code{mlr_sugar}.
}
\examples{
clx_sps("uniform")
}
