% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codomain_helpers.R
\name{codomain_goal}
\alias{codomain_goal}
\title{Determine Goal from Codomain}
\usage{
codomain_goal(codomain)
}
\arguments{
\item{codomain}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
The codomain to inspect.}
}
\value{
\code{character(1)}: "optimize", "learn", or "both".
}
\description{
Infers the search goal based on codomain tags.
}
