% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_codomain.R
\name{codomain_goals}
\alias{codomain_goals}
\title{Determine Goals from Codomain}
\usage{
codomain_goals(codomain)
}
\arguments{
\item{codomain}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
The codomain to inspect.}
}
\value{
\code{character}: subset of \code{"minimize"}, \code{"maximize"}, \code{"learn"}.
Future versions may add more tags.
}
\description{
Lists all search goals based on codomain tags.
}
\keyword{internal}
