% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codomain_helpers.R
\name{codomain_has_learn}
\alias{codomain_has_learn}
\title{Check if Codomain Has Learning Targets}
\usage{
codomain_has_learn(codomain)
}
\arguments{
\item{codomain}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
The codomain to check.}
}
\value{
\code{logical(1)} TRUE if any parameter has "learn" tag.
}
\description{
Tests whether a codomain contains any "learn" tagged parameters.
}
