% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codomain_helpers.R
\name{codomain_learn_ids}
\alias{codomain_learn_ids}
\title{Get Learning Target IDs}
\usage{
codomain_learn_ids(codomain)
}
\arguments{
\item{codomain}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
The codomain to inspect.}
}
\value{
Character vector of learn-tagged target IDs.
}
\description{
Returns IDs of parameters tagged for learning (not optimization).
}
