% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_metrics.R
\name{compute_domain_volume}
\alias{compute_domain_volume}
\title{Compute Domain Volume}
\usage{
compute_domain_volume(search_space)
}
\arguments{
\item{search_space}{(\link[paradox:ParamSet]{paradox::ParamSet}) The search space.}
}
\value{
\code{numeric(1)} The domain volume.
}
\description{
Computes the volume of the search space (product of ranges for numeric
parameters, product of level counts for categorical).
}
\details{
For mixed numeric/categorical spaces, this returns the product of:
\itemize{
\item (upper - lower) for each numeric parameter
\item Number of levels for each categorical parameter
}
}
\keyword{internal}
