% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigurableComponent.R
\name{ConfigurableComponent}
\alias{ConfigurableComponent}
\title{ConfigurableComponent}
\arguments{
\item{id}{(\code{character(1)} | \code{NULL})\cr
Identifier of the object. If left as \code{NULL}, the object does not get an ID.}

\item{param_set}{(\link[paradox:ParamSet]{paradox::ParamSet} | \code{NULL})\cr
Set of hyperparameters / configuration parameters.}

\item{label}{(\code{character(1)})\cr
Label for the object.}

\item{man}{(\code{character(1)})\cr
String in the format \verb{[pkg]::[topic]} pointing to a manual page.}

\item{packages}{(\code{character()})\cr
Package names required by the object.}

\item{additional_configuration}{(\code{character()})\cr
Additional configuration settings.
Anything passed here must be a field or active binding of the object that is
not passed as an initialization argument to the (top-level) constructor.}

\item{additional_phash_input}{(\code{character()})\cr
Names of fields (in \code{self} or \code{private}) whose values should influence the persistent hash (\code{phash}).
Useful for subclasses that expose additional configuration state outside of constructor arguments
or the parameter set.
Note that entries of \code{additional_configuration} are \emph{not} automatically included (as they may be expensive to
compute).
\code{additional_phash_input} should reference the data from which \code{additional_configuration} fields are computed.}

\item{...}{(named \code{any})\cr
Configuration values to set.}

\item{.values}{(named \code{list()})\cr
Alternative way to pass configuration values.}
}
\description{
Base class for objects that unifies configuration of state.
\code{ConfigurableComponent}-objects should always have active bindings that reflect the arguments given to their
constructor.
They also have a \verb{$param_set} field containing a \code{paradox::ParamSet}, which are used for more fine-grained
configuration of the object.
Things that somehow determine how an object can be used (e.g. what kinds of things it is compatible with),
particularly things that can not be changed after constructions should likely be construction arguments;
others should likely be in \verb{$param_set}.

Initialize a ConfigurableComponent.

Helper for print outputs.

Printer.

Set parameter values and fields in one step.
}
\details{
The state of the object should be completely determined by:
\itemize{
\item arguments passed for initialization, all of which must be accessible
as fields or active bindings of the object.
\item param_set configuration parameteres
\item additional fields / active bindings, which must be listed as \code{additional_configuration}.
\item there may be other active bindings, but these should not convey any relevant state.
}
}
\section{Fields}{

\describe{
\item{\code{id}}{(\code{character(1)})
Identifier of the object.
Used in tables, plot and text output.}

\item{\code{label}}{(\code{character(1)})
Label for this object.}

\item{\code{man}}{(\code{character(1)})
String in the format \verb{[pkg]::[topic]} pointing to a manual page.}

\item{\code{packages}}{(\code{character()})
Required packages.}

\item{\code{param_set}}{(\link[paradox:ParamSet]{paradox::ParamSet} | \code{NULL})
Set of hyperparameters.}

\item{\code{hash}}{(\code{character(1)})
Stable hash that includes the parameter values (if present) and everything covered by \code{phash},
but not state.}

\item{\code{phash}}{(\code{character(1)})
Hash that includes id, class, label, man, packages, and the fields named in the
\code{additional_phash_input} construction argument, but not parameter values and no state.}
}}

\concept{ConfigurableComponent}
