% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveExtrapolatorSim.R
\name{curve_objective_builder_default}
\alias{curve_objective_builder_default}
\title{Default Objective Builder for Replay Simulation}
\usage{
curve_objective_builder_default(
  archive,
  learner = lrn("regr.featureless"),
  target_col = NULL,
  sim_id = NULL
)
}
\arguments{
\item{archive}{(\link[bbotk:ArchiveBatch]{bbotk::ArchiveBatch})
Archive with existing evaluations.}

\item{learner}{(\link[mlr3:LearnerRegr]{mlr3::LearnerRegr})
Learner used to fit the proxy objective.}

\item{target_col}{(\code{character(1)} | \code{NULL})
Target column to model. If \code{NULL}, uses first \code{archive$cols_y} entry.}

\item{sim_id}{(\code{integer(1)} | \code{NULL})
Optional simulation id.}
}
\value{
\link{ObjectiveLearner}
}
\description{
Builds an \link{ObjectiveLearner} proxy objective from an archive.
}
