% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_metrics.R
\name{generate_default_grid}
\alias{generate_default_grid}
\title{Generate Default Grid}
\usage{
generate_default_grid(search_space, n = 100L)
}
\arguments{
\item{search_space}{(\link[paradox:ParamSet]{paradox::ParamSet}) The search space to sample from.}

\item{n}{(\code{integer(1)}) Number of grid points. Default 100.}
}
\value{
\code{data.table} with one row per grid point.
}
\description{
Generates a default evaluation grid from a search space using Latin
Hypercube Sampling.
}
\keyword{internal}
