% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_task_lce.R
\name{generate_newdata_lce}
\alias{generate_newdata_lce}
\title{Generate New Data for LCE Prediction}
\usage{
generate_newdata_lce(task, n_evals_new)
}
\arguments{
\item{task}{(\link{TaskLCE})\cr
The training task, used to determine column schema.}

\item{n_evals_new}{(\code{integerish()})\cr
Future evaluation counts to predict at.}
}
\value{
\code{\link[data.table:data.table]{data.table::data.table()}} with the order column set to
\code{n_evals_new}, the target column set to \code{NA}, and other feature columns
set to \code{NA}.
}
\description{
Creates a \code{data.table} suitable for \code{learner$predict_newdata()} containing
future evaluation count values.
}
