% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_al_helpers.R
\name{get_first_target_id}
\alias{get_first_target_id}
\title{Get the first target column ID from a codomain.}
\usage{
get_first_target_id(codomain)
}
\arguments{
\item{codomain}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
A codomain.}
}
\value{
\code{character(1)} first target column name.
}
\description{
Prefers "learn"-tagged targets, then falls back to "minimize"/"maximize".
}
\keyword{internal}
