% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.R
\name{marshal_model.learner_regr_quantile_se_state}
\alias{marshal_model.learner_regr_quantile_se_state}
\title{Marshal Model for LearnerRegrQuantileSE State}
\usage{
\method{marshal_model}{learner_regr_quantile_se_state}(model, inplace = FALSE, ...)
}
\arguments{
\item{model}{(\code{learner_regr_quantile_se_state})\cr
The model to marshal.}

\item{inplace}{(\code{logical(1)})\cr
Whether to marshal in place.}

\item{...}{(any)\cr
Additional arguments passed to \code{\link[mlr3:marshaling]{mlr3::marshal_model()}}.}
}
\value{
Either the original model (if no marshaling was needed) or a
marshaled version with class \code{learner_regr_quantile_se_state_marshaled}.
}
\description{
Marshals the model state of \link{LearnerRegrQuantileSE}, preparing it for
serialization. This marshals the model inside the stored base learner state
(via \link[mlr3:mlr3-package]{mlr3::mlr3}'s \code{marshal_model.learner_state}).
}
\keyword{internal}
