% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_metrics.R
\name{metric_best_y}
\alias{metric_best_y}
\title{Best Y Metric}
\usage{
metric_best_y(archive, surrogate = NULL, target = NULL, ...)
}
\arguments{
\item{archive}{(\link[bbotk:ArchiveBatch]{bbotk::ArchiveBatch})\cr
The archive.}

\item{surrogate}{(\link[mlr3mbo:Surrogate]{mlr3mbo::Surrogate})\cr
Ignored.}

\item{target}{(\code{character(1)} | \code{NULL})\cr
Target column name. If \code{NULL}, uses first codomain target.}

\item{...}{Ignored.}
}
\value{
\code{numeric(1)} minimum y value.
}
\description{
Returns the best (minimum) observed y value in the archive.
}
