% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_metrics.R
\name{metric_integrated_variance}
\alias{metric_integrated_variance}
\title{Integrated Variance Metric}
\usage{
metric_integrated_variance(
  archive,
  surrogate = NULL,
  grid = NULL,
  search_space = NULL,
  ...
)
}
\arguments{
\item{archive}{(\link[bbotk:ArchiveBatch]{bbotk::ArchiveBatch}) The archive (may be used for normalization).}

\item{surrogate}{(\link[mlr3mbo:Surrogate]{mlr3mbo::Surrogate}) The fitted surrogate model.}

\item{grid}{(\code{data.table}) Points at which to evaluate variance. If NULL,
generates a default grid from the search space.}

\item{search_space}{(\link[paradox:ParamSet]{paradox::ParamSet}) Used to generate grid if not provided.}

\item{...}{Ignored.}
}
\value{
\code{numeric(1)} The integrated variance.
}
\description{
Approximates the integral of prediction variance over the domain.
This is the continuous version of mean variance, accounting for
domain volume.
}
\details{
For a uniform grid, this equals mean_variance * domain_volume.
More sophisticated integration (e.g., Monte Carlo) may be added.
}
