% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_metrics.R
\name{metric_simple_regret}
\alias{metric_simple_regret}
\title{Simple Regret Metric}
\usage{
metric_simple_regret(archive, surrogate = NULL, optimum, target = NULL, ...)
}
\arguments{
\item{archive}{(\link[bbotk:ArchiveBatch]{bbotk::ArchiveBatch})\cr
The archive.}

\item{surrogate}{(\link[mlr3mbo:Surrogate]{mlr3mbo::Surrogate})\cr
Ignored.}

\item{optimum}{(\code{numeric(1)})\cr
The known optimal value.}

\item{target}{(\code{character(1)} | \code{NULL})\cr
Target column name.}
}
\description{
Alias for \link{metric_regret}.
}
