% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_metrics.R
\name{metric_worst_y}
\alias{metric_worst_y}
\title{Worst Y Metric}
\usage{
metric_worst_y(archive, surrogate = NULL, target = NULL, ...)
}
\arguments{
\item{archive}{(\link[bbotk:ArchiveBatch]{bbotk::ArchiveBatch}) The archive.}

\item{surrogate}{(\link{SurrogateModel} | \link[mlr3mbo:Surrogate]{mlr3mbo::Surrogate}) Ignored.}

\item{target}{(\code{character(1)}) Target column name.}

\item{...}{Ignored.}
}
\value{
\code{numeric(1)} The maximum y value.
}
\description{
Returns the worst (maximum) observed y value in the archive.
}
