% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrGPfit.R
\name{mlr_learners_regr.gpfit}
\alias{mlr_learners_regr.gpfit}
\alias{LearnerRegrGPfit}
\title{GPfit Regression Learner}
\description{
Gaussian process regression via \code{\link[GPfit:GP_fit]{GPfit::GP_fit()}}.

This learner predicts the posterior mean and returns the square root of the
package's predictive mean squared error as \code{"se"}.

\code{GPfit} assumes that inputs are scaled to the unit hypercube.

Creates a new instance of this learner.
}
