% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerPool.R
\name{mlr_optimizers_pool}
\alias{mlr_optimizers_pool}
\alias{OptimizerPool}
\title{Pool-Based Optimizer}
\description{
Optimizer that generates a pool of candidates and evaluates them all.
This is an extension of \link[bbotk:mlr_optimizers_random_search]{bbotk::OptimizerBatchRandomSearch} that allows
for more flexible candidate generation.

Creates a new instance of this \link[R6:R6Class]{R6} class.
}
\section{Parameters}{

\itemize{
\item \code{candidate_generator} (\code{function})\cr
Function to generate candidate points. Must have signature
\verb{function(search_space, n)} and return a \code{data.table} of n candidates.
Default is \code{\link[=candidate_generator_lhs]{candidate_generator_lhs()}}.
\item \code{pool_size} (\code{integer(1)})\cr
Number of candidates to generate and evaluate. If \code{NULL}, calculates
remaining evaluations from the terminator (requires \link[bbotk:mlr_terminators_evals]{bbotk::TerminatorEvals}).
Default is to use the remaining evaluations from the terminator.
}
}

\examples{
\dontrun{
optimizer <- opt("pool")
optimizer$param_set$set_values(
  candidate_generator = candidate_generator_lhs(),
  pool_size = 100L
)
}

}
