% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_space_samplers.R
\name{mlr_space_samplers}
\alias{mlr_space_samplers}
\title{Dictionary of Space Samplers}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}
\description{
A simple \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} storing objects of class \link{SpaceSampler}.
Each sampler has an associated help page, see \code{mlr_space_samplers_[id]}.

For a more convenient way to retrieve and construct a sampler, see
\code{\link[=clx_sps]{clx_sps()}} and \code{\link[=clx_spss]{clx_spss()}}.
}
\section{Methods}{

See \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}

\examples{
as.data.table(mlr_space_samplers)
clx_sps("uniform")
}
\seealso{
Sugar functions: \code{\link[=clx_sps]{clx_sps()}}, \code{\link[=clx_spss]{clx_spss()}}

Other Dictionary:
\code{\link{mlr_al_distances}}

Other SpaceSampler:
\code{\link{SpaceSampler}},
\code{\link{SpaceSamplerDistance}},
\code{\link{mlr_space_samplers_chain}},
\code{\link{mlr_space_samplers_conditional}},
\code{\link{mlr_space_samplers_gsx}},
\code{\link{mlr_space_samplers_kmeans}},
\code{\link{mlr_space_samplers_kmedoids}},
\code{\link{mlr_space_samplers_lhs}},
\code{\link{mlr_space_samplers_relational_kmeans}},
\code{\link{mlr_space_samplers_sobol}},
\code{\link{mlr_space_samplers_uniform}}
}
\concept{Dictionary}
\concept{SpaceSampler}
