% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceSamplerGSx.R
\name{mlr_space_samplers_gsx}
\alias{mlr_space_samplers_gsx}
\alias{SpaceSamplerGSx}
\title{GSx Space Sampler}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link{SpaceSamplerDistance}.
}
\arguments{
\item{distance}{(\link{ALDistance})\cr
Distance used for the GSx selection rule.}
}
\description{
Greedy-sampling-in-input-space sampler.

Without \code{known_pool}, the first selected point is the pool point minimizing
the sum of squared distances to all other pool points. Subsequent points are
chosen by farthest-first traversal, maximizing the minimum squared distance to
the already-selected or known points.

Creates a new GSx space sampler.
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{clx_sps("gsx")
}\if{html}{\out{</div>}}
}

\seealso{
Other SpaceSampler:
\code{\link{SpaceSampler}},
\code{\link{SpaceSamplerDistance}},
\code{\link{mlr_space_samplers}},
\code{\link{mlr_space_samplers_chain}},
\code{\link{mlr_space_samplers_conditional}},
\code{\link{mlr_space_samplers_kmeans}},
\code{\link{mlr_space_samplers_kmedoids}},
\code{\link{mlr_space_samplers_lhs}},
\code{\link{mlr_space_samplers_relational_kmeans}},
\code{\link{mlr_space_samplers_sobol}},
\code{\link{mlr_space_samplers_uniform}}
}
\concept{SpaceSampler}
