% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceSamplerLhs.R
\name{mlr_space_samplers_lhs}
\alias{mlr_space_samplers_lhs}
\alias{SpaceSamplerLhs}
\title{LHS Space Sampler}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link{SpaceSampler}.
}
\description{
Sampler that draws Latin hypercube designs from a search space.

This sampler only samples from the search space directly and errors when a
\code{pool} is supplied.

Creates a new LHS space sampler.
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{clx_sps("lhs")
}\if{html}{\out{</div>}}
}

\seealso{
Other SpaceSampler:
\code{\link{SpaceSampler}},
\code{\link{SpaceSamplerDistance}},
\code{\link{mlr_space_samplers}},
\code{\link{mlr_space_samplers_chain}},
\code{\link{mlr_space_samplers_conditional}},
\code{\link{mlr_space_samplers_gsx}},
\code{\link{mlr_space_samplers_kmeans}},
\code{\link{mlr_space_samplers_kmedoids}},
\code{\link{mlr_space_samplers_relational_kmeans}},
\code{\link{mlr_space_samplers_sobol}},
\code{\link{mlr_space_samplers_uniform}}
}
\concept{SpaceSampler}
