% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerSearchAbstract.R
\name{OptimizerSearchAbstract}
\alias{OptimizerSearchAbstract}
\title{Abstract Base Class for Search-Compatible Optimizers}
\description{
Extends \link[bbotk:OptimizerBatch]{bbotk::OptimizerBatch} with support for \link{SearchInstance} in
addition to \link[bbotk:OptimInstance]{bbotk::OptimInstance}.
Optimizers that should work with both \link[bbotk:OptimInstance]{bbotk::OptimInstance} and
\link{SearchInstance} should inherit from this class instead of
\link[bbotk:OptimizerBatch]{bbotk::OptimizerBatch}.
}
\details{
The default \code{.assign_result()} in \link[bbotk:OptimizerBatch]{bbotk::OptimizerBatch}
assumes an \link[bbotk:OptimInstance]{bbotk::OptimInstance} and errors on a \link{SearchInstance}.
This class overrides \code{.assign_result()} to handle both: it delegates to
\code{\link[bbotk:assign_result_default]{bbotk::assign_result_default()}} for \link[bbotk:OptimInstance]{bbotk::OptimInstance} and is a no-op
for \link{SearchInstance} (which has no result slot).
}
