% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_param_set.R
\name{param_set_topo_ids}
\alias{param_set_topo_ids}
\title{Topologically sort parameter IDs by dependency order}
\usage{
param_set_topo_ids(param_set)
}
\arguments{
\item{param_set}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Parameter set to sort.}
}
\value{
\code{character()} of parameter IDs in topological order.
}
\description{
"Parent" parameters that other parameters depend on are sorted first.
}
\keyword{internal}
