% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_objective.R
\name{param_set_valid_mask_dt}
\alias{param_set_valid_mask_dt}
\title{Get \code{logical(nrow(dt))} mask of valid configurations in a data.table}
\usage{
param_set_valid_mask_dt(dt, param_set)
}
\arguments{
\item{dt}{(\code{data.table})\cr
\code{data.table} to validate.}

\item{param_set}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
\code{ParamSet} describing the space to validate against.}
}
\description{
Returns a \code{logical(nrow(dt))} indicating, for each row, whether the configuration is valid in the \code{ParamSet}.
}
\keyword{internal}
