% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchInstance.R
\name{search_instance}
\alias{search_instance}
\title{Create Search Instance}
\usage{
search_instance(objective, terminator, search_space = NULL, ...)
}
\arguments{
\item{objective}{(\link[bbotk:Objective]{bbotk::Objective})\cr
The objective to evaluate.}

\item{terminator}{(\link[bbotk:Terminator]{bbotk::Terminator})\cr
Termination criterion.}

\item{search_space}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Optional restricted search space.}

\item{...}{Additional arguments passed to SearchInstance$new().}
}
\value{
A \link{SearchInstance} object.
}
\description{
Convenience constructor for \link{SearchInstance}.
}
\examples{
\dontrun{
# Create instance with evaluation budget
terminator <- bbotk::TerminatorEvals$new()
terminator$param_set$set_values(n_evals = 100L, k = 0L)
instance <- search_instance(objective, terminator = terminator)
}

}
