% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchInstance.R
\name{search_terminated_error}
\alias{search_terminated_error}
\title{Search Terminated Error}
\usage{
search_terminated_error(search_instance)
}
\arguments{
\item{search_instance}{(\link{SearchInstance})\cr
The SearchInstance that terminated.}
}
\value{
A condition object with class \code{search_terminated_error}
(inherits from \code{terminated_error}, \code{error}, \code{condition}).
}
\description{
Creates a condition indicating that a SearchInstance has terminated.
Inherits from bbotk's \code{terminated_error}.
}
