% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_scaling.R
\name{standardize_apply}
\alias{standardize_apply}
\title{Apply standardization to new data.}
\usage{
standardize_apply(x, scaler)
}
\arguments{
\item{x}{(\code{matrix})\cr
Numeric matrix with same number of columns as original pool.}

\item{scaler}{(\code{list})\cr
List returned by \code{\link[=standardize_features]{standardize_features()}}.}
}
\value{
Scaled \code{matrix}.
}
\description{
Apply standardization to new data.
}
\keyword{internal}
