% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marshal.R
\name{unmarshal_model.learner_regr_bootstrap_se_state_marshaled}
\alias{unmarshal_model.learner_regr_bootstrap_se_state_marshaled}
\title{Unmarshal Model for LearnerRegrBootstrapSE State}
\usage{
\method{unmarshal_model}{learner_regr_bootstrap_se_state_marshaled}(model, inplace = FALSE, ...)
}
\arguments{
\item{model}{(\code{learner_regr_bootstrap_se_state_marshaled})\cr
The marshaled model to unmarshal.}

\item{inplace}{(\code{logical(1)})\cr
Whether to unmarshal in place.}

\item{...}{(any)\cr
Additional arguments passed to \code{\link[mlr3:unmarshal_model]{mlr3::unmarshal_model()}}.}
}
\value{
The unmarshaled model state.
}
\description{
Unmarshals the model state of \link{LearnerRegrBootstrapSE}, restoring it after
deserialization.
}
\keyword{internal}
