% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcqFunctionDistGSx.R
\name{AcqFunctionDistGSx}
\alias{AcqFunctionDistGSx}
\title{Distance-Aware GSx Acquisition Function}
\arguments{
\item{surrogate}{(\code{NULL} | \link[mlr3mbo:Surrogate]{mlr3mbo::Surrogate})\cr
Optional surrogate. Only used to derive the acquisition domain and
archive reference points.}

\item{al_distance}{(\link{ALDistance} | \code{NULL})\cr
Distance used for input-space scores.}
}
\description{
Greedy-sampling-in-input-space acquisition function using an \link{ALDistance}.

The score is the distance to the nearest already evaluated reference point.

Creates a new distance-aware GSx acquisition function.
}
\seealso{
Other Acquisition Function: 
\code{\link{AcqFunctionDist}},
\code{\link{AcqFunctionDistIDEAL}},
\code{\link{AcqFunctionDistIGS}},
\code{\link{AcqFunctionGSy}},
\code{\link{SurrogateNull}}
}
\concept{Acquisition Function}
