% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcqFunctionDistIGS.R
\name{AcqFunctionDistIGS}
\alias{AcqFunctionDistIGS}
\title{Distance-Aware iGS Acquisition Function}
\arguments{
\item{surrogate}{(\code{NULL} | \link[mlr3mbo:SurrogateLearner]{mlr3mbo::SurrogateLearner})\cr
Surrogate used for mean predictions.}

\item{al_distance}{(\link{ALDistance} | \code{NULL})\cr
Distance used for input-space scores.}
}
\description{
Improved greedy-sampling acquisition function using an \link{ALDistance}.

The score is the minimum, over already evaluated reference points, of the
product of input-space distance and output-space distance.

Creates a new distance-aware iGS acquisition function.
}
\seealso{
Other Acquisition Function: 
\code{\link{AcqFunctionDist}},
\code{\link{AcqFunctionDistGSx}},
\code{\link{AcqFunctionDistIDEAL}},
\code{\link{AcqFunctionGSy}},
\code{\link{SurrogateNull}}
}
\concept{Acquisition Function}
