% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{celecx-package}
\alias{celecx}
\alias{celecx-package}
\title{celecx: Computer Experiment LEarning Curve eXtrapolation}
\description{
Under development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://celecx.mlr-org.com/}
  \item \url{https://github.com/mlr-org/celecx}
  \item Report bugs at \url{https://github.com/mlr-org/celecx/issues}
}

}
\author{
\strong{Maintainer}: Martin Binder \email{mlr.developer@mb706.com}

}
