% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_space_samplers.R
\name{clx_spss}
\alias{clx_spss}
\title{Syntactic Sugar Space Samplers Construction}
\usage{
clx_spss(.keys, ...)
}
\arguments{
\item{.keys}{(\code{character()})\cr
Keys passed to \link{mlr_space_samplers}.}

\item{...}{(named \code{list()})\cr
Named arguments passed to the constructor, to the parameter set, or to
public fields. See \code{\link[mlr3misc:dictionary_sugar_get]{mlr3misc::dictionary_sugar_mget()}} for details.}
}
\value{
Named \code{list} of \link{SpaceSampler} objects.
}
\description{
Retrieve multiple space samplers from \link{mlr_space_samplers}.
}
\examples{
clx_spss(c("uniform", "gsx"))
}
