% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchInstance.R
\name{ContextSearch}
\alias{ContextSearch}
\title{Context for Search Instance}
\arguments{
\item{inst}{(\link{SearchInstance}).}

\item{optimizer}{(\link[bbotk:Optimizer]{bbotk::Optimizer} | \code{NULL}).}
}
\description{
Context object that allows callbacks to access and modify data during
search operations on a \link{SearchInstance}.

Creates a new instance of this \link[R6:R6Class]{R6} class.
}
\details{
Similar to bbotk's \code{ContextBatch}, but for \link{SearchInstance} instead of
\link[bbotk:OptimInstanceBatch]{bbotk::OptimInstanceBatch}.
}
\section{Fields}{

\describe{
\item{\code{instance}}{(\link{SearchInstance}).}

\item{\code{optimizer}}{(\link[bbotk:Optimizer]{bbotk::Optimizer} | \code{NULL}).}

\item{\code{xdt}}{(\link[data.table:data.table]{data.table::data.table})\cr
The points of the latest batch in \code{instance$eval_batch()}.}
}}

