% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_al_distances.R
\docType{data}
\name{mlr_al_distances}
\alias{mlr_al_distances}
\title{Dictionary of Active Learning Distances}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}
\description{
A simple \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} storing objects of class \link{ALDistance}.
Each distance has an associated help page, see \code{mlr_al_distances_[id]}.

For a more convenient way to retrieve and construct a distance, see
\code{\link[=clx_ald]{clx_ald()}} and \code{\link[=clx_alds]{clx_alds()}}.
}
\section{Methods}{

See \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}

\examples{
as.data.table(mlr_al_distances)
clx_ald("standardize")
}
\seealso{
Sugar functions: \code{\link[=clx_ald]{clx_ald()}}, \code{\link[=clx_alds]{clx_alds()}}

Other Dictionary: 
\code{\link{mlr_al_objectives}},
\code{\link{mlr_space_samplers}}

Other ALDistance: 
\code{\link{ALDistance}},
\code{\link{ALDistanceGeometry}},
\code{\link{mlr_al_distances_affine}},
\code{\link{mlr_al_distances_gower}},
\code{\link{mlr_al_distances_standardize}}
}
\concept{ALDistance}
\concept{Dictionary}
\keyword{datasets}
