% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrTGP.R
\name{mlr_learners_regr.tgp}
\alias{mlr_learners_regr.tgp}
\alias{LearnerRegrTGP}
\title{tgp Regression Learner}
\description{
Bayesian Gaussian process regression via \code{\link[tgp:btgp]{tgp::bgp()}}.

The learner uses the package's prediction routine \code{\link[stats:predict]{stats::predict()}} for
fitted \code{"tgp"} objects and returns the posterior predictive mean together
with the square root of the predictive variance.

Creates a new instance of this learner.
}
