% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceSamplerConditional.R
\name{mlr_space_samplers_conditional}
\alias{mlr_space_samplers_conditional}
\alias{SpaceSamplerConditional}
\title{Conditional Space Sampler}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link{SpaceSampler}.
}
\arguments{
\item{on_discrete}{(\link{SpaceSampler})\cr
Sampler used when a \code{pool} is supplied.}

\item{on_continuous}{(\link{SpaceSampler})\cr
Sampler used when \code{pool} is \code{NULL}.}
}
\description{
Wrapper sampler that delegates to one sampler for finite candidate pools and
another sampler for direct search-space sampling.

Creates a new conditional space sampler.
}
\section{Fields}{

\describe{
\item{\code{on_discrete}}{(\link{SpaceSampler})
Sampler used when a \code{pool} is supplied.}

\item{\code{on_continuous}}{(\link{SpaceSampler})
Sampler used when \code{pool} is \code{NULL}.}
}}

\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{clx_sps("conditional")
}\if{html}{\out{</div>}}
}

\seealso{
Other SpaceSampler: 
\code{\link{SpaceSampler}},
\code{\link{SpaceSamplerDistance}},
\code{\link{mlr_space_samplers}},
\code{\link{mlr_space_samplers_chain}},
\code{\link{mlr_space_samplers_gsx}},
\code{\link{mlr_space_samplers_kmeans}},
\code{\link{mlr_space_samplers_kmedoids}},
\code{\link{mlr_space_samplers_lhs}},
\code{\link{mlr_space_samplers_relational_kmeans}},
\code{\link{mlr_space_samplers_sobol}},
\code{\link{mlr_space_samplers_uniform}}
}
\concept{SpaceSampler}
