% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceSamplerKMeans.R
\name{mlr_space_samplers_kmeans}
\alias{mlr_space_samplers_kmeans}
\alias{SpaceSamplerKMeans}
\title{K-Means Space Sampler}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link{SpaceSamplerDistance}.
}
\arguments{
\item{distance}{(\link{ALDistance})\cr
Distance used to construct the fitted geometry.}
}
\description{
K-means-based pool sampler.

The wrapped distance must expose a numeric reference embedding or numeric
reference columns after \code{set_reference_points()}.

Creates a new K-means space sampler.
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{clx_sps("kmeans")
}\if{html}{\out{</div>}}
}

\seealso{
Other SpaceSampler: 
\code{\link{SpaceSampler}},
\code{\link{SpaceSamplerDistance}},
\code{\link{mlr_space_samplers}},
\code{\link{mlr_space_samplers_chain}},
\code{\link{mlr_space_samplers_conditional}},
\code{\link{mlr_space_samplers_gsx}},
\code{\link{mlr_space_samplers_kmedoids}},
\code{\link{mlr_space_samplers_lhs}},
\code{\link{mlr_space_samplers_relational_kmeans}},
\code{\link{mlr_space_samplers_sobol}},
\code{\link{mlr_space_samplers_uniform}}
}
\concept{SpaceSampler}
