% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceSamplerKMedoids.R
\name{mlr_space_samplers_kmedoids}
\alias{mlr_space_samplers_kmedoids}
\alias{SpaceSamplerKMedoids}
\title{K-Medoids Space Sampler}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link{SpaceSamplerDistance}.
}
\arguments{
\item{distance}{(\link{ALDistance})\cr
Distance used for the medoid objective.}
}
\description{
K-medoids-style pool sampler based on an arbitrary \link{ALDistance}.

Medoids are selected from the candidate pool using a greedy build step
followed by swap-based local improvement. The optimization target is the sum
of distances to the nearest selected medoid.

If \code{known_pool} is given, those points are treated as fixed medoids and the
selected pool points are optimized to complement them.

Creates a new k-medoids space sampler.
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{clx_sps("kmedoids")
}\if{html}{\out{</div>}}
}

\seealso{
Other SpaceSampler: 
\code{\link{SpaceSampler}},
\code{\link{SpaceSamplerDistance}},
\code{\link{mlr_space_samplers}},
\code{\link{mlr_space_samplers_chain}},
\code{\link{mlr_space_samplers_conditional}},
\code{\link{mlr_space_samplers_gsx}},
\code{\link{mlr_space_samplers_kmeans}},
\code{\link{mlr_space_samplers_lhs}},
\code{\link{mlr_space_samplers_relational_kmeans}},
\code{\link{mlr_space_samplers_sobol}},
\code{\link{mlr_space_samplers_uniform}}
}
\concept{SpaceSampler}
