% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceSamplerRelationalKMeans.R
\name{mlr_space_samplers_relational_kmeans}
\alias{mlr_space_samplers_relational_kmeans}
\alias{SpaceSamplerRelationalKMeans}
\title{Relational K-Means Space Sampler}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link{SpaceSamplerDistance}.
}
\arguments{
\item{distance}{(\link{ALDistance})\cr
Distance used to construct the pairwise dissimilarity matrix.}
}
\description{
K-means-style pool sampler driven directly by the pairwise distance matrix.

This sampler clusters the fitted pool with the relational k-means objective.
For Euclidean distances it is equivalent to ordinary k-means in an unknown
embedding. For non-Euclidean distances it remains a heuristic.

If \code{known_pool} is given, those points influence the implicit centroids but
are not eligible for selection.

Creates a new relational k-means space sampler.
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{clx_sps("relational_kmeans")
}\if{html}{\out{</div>}}
}

\seealso{
Other SpaceSampler: 
\code{\link{SpaceSampler}},
\code{\link{SpaceSamplerDistance}},
\code{\link{mlr_space_samplers}},
\code{\link{mlr_space_samplers_chain}},
\code{\link{mlr_space_samplers_conditional}},
\code{\link{mlr_space_samplers_gsx}},
\code{\link{mlr_space_samplers_kmeans}},
\code{\link{mlr_space_samplers_kmedoids}},
\code{\link{mlr_space_samplers_lhs}},
\code{\link{mlr_space_samplers_sobol}},
\code{\link{mlr_space_samplers_uniform}}
}
\concept{SpaceSampler}
