% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceSamplerUniform.R
\name{mlr_space_samplers_uniform}
\alias{mlr_space_samplers_uniform}
\alias{SpaceSamplerUniform}
\title{Uniform Space Sampler}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link{SpaceSampler}.
}
\description{
Sampler that draws uniformly from the search space or uniformly without
replacement from a given pool.

Creates a new uniform space sampler.
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{clx_sps("uniform")
}\if{html}{\out{</div>}}
}

\seealso{
Other SpaceSampler: 
\code{\link{SpaceSampler}},
\code{\link{SpaceSamplerDistance}},
\code{\link{mlr_space_samplers}},
\code{\link{mlr_space_samplers_chain}},
\code{\link{mlr_space_samplers_conditional}},
\code{\link{mlr_space_samplers_gsx}},
\code{\link{mlr_space_samplers_kmeans}},
\code{\link{mlr_space_samplers_kmedoids}},
\code{\link{mlr_space_samplers_lhs}},
\code{\link{mlr_space_samplers_relational_kmeans}},
\code{\link{mlr_space_samplers_sobol}}
}
\concept{SpaceSampler}
