% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_generators.R
\name{sample_local}
\alias{sample_local}
\title{Sample in Local Region}
\usage{
sample_local(center, n, radius, search_space)
}
\arguments{
\item{center}{(\code{data.table} or \code{list}) Center point.}

\item{n}{(\code{integer(1)}) Number of points to generate.}

\item{radius}{(\code{numeric(1)}) Radius as fraction of domain range.}

\item{search_space}{(\link[paradox:ParamSet]{paradox::ParamSet}) The search space.}
}
\value{
\code{data.table} of n points.
}
\description{
Generates random points in a hypercube around a center point.
}
\keyword{internal}
