% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceSamplerDistance.R
\name{SpaceSamplerDistance}
\alias{SpaceSamplerDistance}
\title{Distance-Based Space Sampler}
\arguments{
\item{id}{(\code{character(1)})\cr
Identifier of the sampler.}

\item{distance}{(\link{ALDistance})\cr
Distance object used by the sampler.}
}
\description{
Abstract base class for pool samplers that are driven by an \link{ALDistance}.

The wrapped distance object is stored directly and its \link[paradox:ParamSet]{paradox::ParamSet} is
forwarded as the sampler parameter set.

Creates a new distance-based space sampler.
}
\section{Fields}{

\describe{
\item{\code{distance}}{(\link{ALDistance})
Wrapped distance object.}
}}

\seealso{
Other SpaceSampler: 
\code{\link{SpaceSampler}},
\code{\link{mlr_space_samplers}},
\code{\link{mlr_space_samplers_chain}},
\code{\link{mlr_space_samplers_conditional}},
\code{\link{mlr_space_samplers_gsx}},
\code{\link{mlr_space_samplers_kmeans}},
\code{\link{mlr_space_samplers_kmedoids}},
\code{\link{mlr_space_samplers_lhs}},
\code{\link{mlr_space_samplers_relational_kmeans}},
\code{\link{mlr_space_samplers_sobol}},
\code{\link{mlr_space_samplers_uniform}}
}
\concept{SpaceSampler}
